Fix some compiler warnings.
* src/prt.cc (main): remove useless cast.
* src/sf-cdc.cc (inlist): delete this unused function.
* src/sid.h (sid): add method get_release().
* src/sid.cc (sid::get_release): implement this method.
* src/sf-get2.cc (sccs_file::sid_matches): instead of casting value to
type release, call sid::get_release() to extract the release number
from SIDs.

+++ src/prt.cc
@@ -148,7 +148,7 @@ main(int argc, char **argv)
 	      return 2;
 	    }
 
-	  last_cutoff_type = (int)c;
+	  last_cutoff_type = c;
 
 	  sccs_date date = sccs_date(opts.getarg());
 	  if (!date.valid())
